requirements.txt
# Core dependencies
opencv-python>=4.8.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
mediapipe>=0.10.0

# Face recognition
facenet-pytorch>=2.5.0

# Data processing
scikit-learn>=1.3.0
scipy>=1.11.0
Pillow>=10.0.0

# Utilities
imutils>=0.5.4
matplotlib>=3.8.0
tqdm>=4.66.0

# Logging and config
pyyaml>=6.0.0
colorama>=0.4.6

# Optional: for better performance
# dlib>=19.24.0  # Uncomment if you want to use dlib features

# Note: Using >= instead of == for better compatibility
# This allows pip to install the latest compatible versions